Moved lib portion of the cargo into a lib directory. Made the lib.rs easy to work with for both rust and C inter-op languages. Fixed issue with file-loader dealing with exe path when it really shouldn't have. Updated readme.
